Move TagNotFoundException to Tag package
- Exception was wrongly placed under Person package
