+++ src/main/java/seedu/address/model/Model.java
@@ -6,7 +6,7 @@ import javafx.collections.ObservableList;
 import seedu.address.model.person.ReadOnlyPerson;
 import seedu.address.model.person.exceptions.DuplicatePersonException;
 import seedu.address.model.person.exceptions.PersonNotFoundException;
-import seedu.address.model.person.exceptions.TagNotFoundException;
+import seedu.address.model.tag.TagNotFoundException;
 import seedu.address.model.tag.Tag;
 
 /**
