+++ src/main/java/seedu/address/model/ModelManager.java
@@ -15,7 +15,7 @@ import seedu.address.commons.events.model.AddressBookChangedEvent;
 import seedu.address.model.person.ReadOnlyPerson;
 import seedu.address.model.person.exceptions.DuplicatePersonException;
 import seedu.address.model.person.exceptions.PersonNotFoundException;
-import seedu.address.model.person.exceptions.TagNotFoundException;
+import seedu.address.model.tag.TagNotFoundException;
 import seedu.address.model.tag.Tag;
 
 /**
