+++ src/main/java/seedu/address/model/tag/TagNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.tag;
 
 /**
  * Signals that the operation is unable to find the specified tag.
